# Parallel kernel compilation