Refactors BEAM's parallel compilation to enable generic parallel compilation. Next step is to actually do it