@marcellofuschi 2024-01-07
I added parallel kernel compilation to `realize`.

`WINO=1 python3 examples/hlb_cifar10.py` performance on M1 Pro:

Before:
```text
shuffling training dataset in 1011.06 ms (0)
  0 38269.06 ms run, 38268.71 ms python,    0.34 ms CL, 1199.83 loss, 0.000015 LR, 0.84 GB used,     12.94 GFLOPS,    495.18 GOPS
  1 29546.97 ms run, 29546.03 ms python,    0.94 ms CL, 1199.65 loss, 0.000030 LR, 4.34 GB used,     16.76 GFLOPS,    495.08 GOPS
  2  492.04 ms run,    7.23 ms python,  484.81 ms CL, 1191.60 loss, 0.000045 LR, 4.35 GB used,   1006.18 GFLOPS,    495.08 GOPS
  3  418.31 ms run,    8.61 ms python,  409.70 ms CL, 1173.59 loss, 0.000060 LR, 4.36 GB used,   1183.50 GFLOPS,    495.08 GOPS
  4  415.54 ms run,    8.24 ms python,  407.30 ms CL, 1160.32 loss, 0.000075 LR, 4.36 GB used,   1191.40 GFLOPS,    495.08 GOPS
...
```

After (with `PARALLEL_COMPILATION=1`):
```text
shuffling training dataset in 879.53 ms (0)
  0 13002.57 ms run, 13002.24 ms python,    0.33 ms CL, 1199.83 loss, 0.000015 LR, 0.84 GB used,     38.11 GFLOPS,    495.58 GOPS
  1 11115.04 ms run, 11114.08 ms python,    0.96 ms CL, 1199.65 loss, 0.000030 LR, 4.36 GB used,     44.53 GFLOPS,    494.96 GOPS
  2  430.68 ms run,    5.67 ms python,  425.02 ms CL, 1191.60 loss, 0.000045 LR, 4.37 GB used,   1149.24 GFLOPS,    494.96 GOPS
  3  414.49 ms run,    5.17 ms python,  409.32 ms CL, 1173.59 loss, 0.000060 LR, 4.38 GB used,   1194.14 GFLOPS,    494.96 GOPS
  4  412.59 ms run,    5.96 ms python,  406.63 ms CL, 1160.32 loss, 0.000075 LR, 4.38 GB used,   1199.64 GFLOPS,    494.96 GOPS
...
```