@geohot 2024-01-19
Seems very complex looking and is +38 lines. What's the minimal way to do this? Is all this complexity needed?