@airpods69 2024-02-01
Seems like the bounty got closed 2 days ago and I'll open it up again. 
Gonna focus just on the factoring and not speeding the thing up or anything like that. 


If the previous author shows up again then I'm down to discuss what all they were trying to do and we can collaborate.